// frontend/lib/api/model/setResponse.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Workout API
 * A workout tracking API built with FastAPI
 * OpenAPI spec version: 1.0.0
 */
import type { SetResponseNoteText } from "./setResponseNoteText";

/**
 * Schema for set responses.
 */
export interface SetResponse {
  clean_reps: number;
  created_at: string;
  exercise_id: number;
  forced_reps: number;
  id: number;
  note_text?: SetResponseNoteText;
  updated_at: string;
  weight: number;
}